Improves battery indicator with charging status (#51)
* Refactors settings usage for logging and notifications

Changes the implementation to access settings for logging,
notifications, and colors through the HeadsetControl class
properties instead of global variables. This improves code
organization and encapsulation.

* Improves battery indicator with charging status

Adds a "+" symbol to the battery percentage indicator when the
device is charging to provide clearer feedback to the user.

Also, adds battery status property to store the charging status.